Add async test method
The current test framework can not be run under an async test.
This adds an alternate method for processing a command that can be used instead.
