+++ src/test/validator.rs
@@ -1,119 +1,5 @@
 use crate::aggregate::Aggregate;
 
-/// A framework for rigorously testing the aggregate logic, one of the ***most important***
-/// parts of any DDD system.
-///
-pub struct TestFramework<A: Aggregate> {
-    service: A::Services,
-}
-
-impl<A: Aggregate> TestFramework<A> {
-    /// Create a test framework using the provided service.
-    pub fn with(service: A::Services) -> Self {
-        Self { service }
-    }
-}
-
-impl<A> TestFramework<A>
-where
-    A: Aggregate,
-{
-    /// Initiates an aggregate test with no previous events.
-    ///
-    /// ```
-    /// # use cqrs_es::doc::{MyAggregate, MyService};
-    /// use cqrs_es::test::TestFramework;
-    ///
-    /// let executor = TestFramework::<MyAggregate>::with(MyService)
-    ///     .given_no_previous_events();
-    /// ```
-    #[must_use]
-    pub fn given_no_previous_events(self) -> AggregateTestExecutor<A> {
-        AggregateTestExecutor {
-            events: Vec::new(),
-            service: self.service,
-        }
-    }
-    /// Initiates an aggregate test with a collection of previous events.
-    ///
-    /// ```
-    /// # use cqrs_es::doc::{MyAggregate, MyEvents, MyService};
-    /// use cqrs_es::test::TestFramework;
-    ///
-    /// let executor = TestFramework::<MyAggregate>::with(MyService)
-    ///     .given(vec![MyEvents::SomethingWasDone]);
-    /// ```
-    #[must_use]
-    pub fn given(self, events: Vec<A::Event>) -> AggregateTestExecutor<A> {
-        AggregateTestExecutor {
-            events,
-            service: self.service,
-        }
-    }
-}
-
-/// Holds the initial event state of an aggregate and accepts a command.
-pub struct AggregateTestExecutor<A>
-where
-    A: Aggregate,
-{
-    events: Vec<A::Event>,
-    service: A::Services,
-}
-
-impl<A> AggregateTestExecutor<A>
-where
-    A: Aggregate,
-{
-    /// Consumes a command and using the state details previously passed provides a validator object
-    /// to test against.
-    ///
-    /// ```
-    /// # use cqrs_es::doc::{MyAggregate, MyCommands, MyService};
-    /// use cqrs_es::test::TestFramework;
-    ///
-    /// let executor = TestFramework::<MyAggregate>::with(MyService)
-    ///     .given_no_previous_events();
-    ///
-    /// let validator = executor.when(MyCommands::DoSomething);
-    /// ```
-    pub fn when(self, command: A::Command) -> AggregateResultValidator<A> {
-        let result = when::<A>(self.events, command, self.service);
-        AggregateResultValidator { result }
-    }
-
-    /// Adds additional events to an aggregate test.
-    ///
-    /// ```
-    /// # use cqrs_es::doc::{MyAggregate, MyEvents, MyService};
-    /// use cqrs_es::test::TestFramework;
-    ///
-    /// let executor = TestFramework::<MyAggregate>::with(MyService)
-    ///     .given(vec![MyEvents::SomethingWasDone])
-    ///     .and(vec![MyEvents::SomethingElseWasDone]);
-    /// ```
-    #[must_use]
-    pub fn and(self, new_events: Vec<A::Event>) -> Self {
-        let mut events = self.events;
-        events.extend(new_events);
-        let service = self.service;
-        AggregateTestExecutor { events, service }
-    }
-}
-
-#[tokio::main(flavor = "current_thread")]
-async fn when<A: Aggregate>(
-    events: Vec<A::Event>,
-    command: A::Command,
-    service: A::Services,
-) -> Result<Vec<A::Event>, A::Error> {
-    let mut aggregate = A::default();
-    for event in events {
-        aggregate.apply(event);
-    }
-    aggregate.handle(command, &service).await
-}
-
 /// Validation object for the `TestFramework` package.
 pub struct AggregateResultValidator<A>
 where
@@ -184,6 +70,10 @@ impl<A: Aggregate> AggregateResultValidator<A> {
     /// ```
     pub fn inspect_result(self) -> Result<Vec<A::Event>, A::Error> {
         self.result
+    }
+
+    pub(crate) fn new(result: Result<Vec<A::Event>, A::Error>) -> Self {
+        Self { result }
     }
 }
 impl<A> AggregateResultValidator<A>
